Initial commit
This repo contains build instructions for a Docker image. The Docker
image is used in CI for the Calliope project. Each commit will trigger
an automatic build on quay.io for the image
https://quay.io/repository/ssssam/calliope-testsuite

Previously these files were kept in the main Calliope repo, but quay.io
triggered builds on every single push which results in unexpected changes
to the hosted image.